refactor(observability): move reason redaction down to the shared layer
`sanitize_reason`/`redact_token` scrub an upstream `e.to_string()` of paths,
URLs, host:port, emails and credential fragments before it is shown. They lived
in an L2 module, so an L1 store that needs to persist a reason could not reach
them and would have to rely on every current and future board keeping secrets
out of its error text - an invariant enforced nowhere.

Moved verbatim to `observability` (L0), which the layers that need it can all
reach downward. `commands::support` now imports from there instead of reaching
through the autopilot helpers.
