+++ apps/desktop/src-tauri/src/autopilot_helpers/mod.rs
@@ -183,149 +183,9 @@ fn resolve_watched_companies(
     map
 }
 
-/// Max length of a single board's sanitized reason in the user-visible step log.
-/// Diagnostics are a hint, not a full error dump — keep them short and bounded.
-const MAX_REASON_LEN: usize = 200;
-
-/// Redact a board's raw error/skip text before it reaches the user-visible step
-/// log. The text can originate from an upstream `e.to_string()` and may carry
-/// absolute filesystem paths, full URLs, or request internals — emitting those raw
-/// violates the repo path-privacy rule. This collapses each such fragment to a
-/// neutral placeholder, normalises whitespace, and caps the length.
-///
-/// Deliberately conservative: it errs toward over-redaction (a placeholder is
-/// always safe) and keeps the high-level message (e.g. `"429 Too Many Requests"`,
-/// `"needs-login"`, `"network timeout"`) intact. Pure + unit-testable.
-fn sanitize_reason(reason: &str) -> String {
-    let mut out = String::with_capacity(reason.len());
-
-    for token in reason.split_whitespace() {
-        if !out.is_empty() {
-            out.push(' ');
-        }
-        out.push_str(&redact_token(token));
-    }
-
-    if out.chars().count() > MAX_REASON_LEN {
-        let truncated: String = out.chars().take(MAX_REASON_LEN).collect();
-        out = format!("{truncated}…");
-    }
-    out
-}
-
-/// Classify a single whitespace-delimited token and replace it with a neutral
-/// placeholder when it looks like a path / URL / request internal; otherwise keep
-/// it verbatim. Whitespace-token granularity keeps the surrounding human message
-/// (`"failed: <path>"` → `"failed: <path-redacted>"`) readable.
-pub(crate) fn redact_token(token: &str) -> String {
-    // Strip surrounding punctuation (quotes, parens, braces, backticks, trailing
-    // `.,:;|`) so a token like `(C:\Users\x)`, `` `https://…` ``, or `{path}` still
-    // matches, then re-attach it.
-    let trimmed = token.trim_matches(|c: char| {
-        matches!(
-            c,
-            '"' | '\''
-                | '`'
-                | '('
-                | ')'
-                | '['
-                | ']'
-                | '{'
-                | '}'
-                | '<'
-                | '>'
-                | '|'
-                | ','
-                | ';'
-                | ':'
-        )
-    });
-
-    let is_url = trimmed.contains("://");
-    // Windows absolute path: drive letter + `:\` or `:/` (e.g. `C:\Users\…`).
-    let mut chars = trimmed.chars();
-    let is_windows_path = matches!(
-        (chars.next(), chars.next(), chars.next()),
-        (Some(c), Some(':'), Some('\\' | '/')) if c.is_ascii_alphabetic()
-    );
-    // Unix absolute path: starts with `/` and has a further separator (a lone `/`
-    // or a fraction like `1/2` is not a path).
-    let is_unix_path = trimmed.starts_with('/') && trimmed[1..].contains('/');
-    // Drive-less home/user path: a fragment like `Users\alice\…` or `home/alice/…`
-    // that lost its drive letter / leading `/` (common in unwound error chains).
-    // Lowercased substring match catches both separators and any case.
-    let lower = trimmed.to_ascii_lowercase();
-    let is_homeish_path =
-        lower.contains("users\\") || lower.contains("users/") || lower.contains("home/");
-
-    // Standalone credential assignment (`app_key=…`, `token=…`, `secret=…`) emitted
-    // OUTSIDE a full `://` URL. The `marker=` shape (the `=` makes it an assignment)
-    // is what flags it — matching the bare word would over-redact `keyword` / a
-    // prose "token". The `://` branch below runs first, so a full
-    // `https://…?app_key=…` still collapses to `<url-redacted>`, not this.
-    // The JSON-field shape (`"api_key":"value"` → brace/quote-trimmed to
-    // `api_key":"value`) is also matched via the `key":` / `token":` sub-strings
-    // so structured log lines (e.g. `{"api_key":"sk-…"}`) don't bypass redaction.
-    let is_credential = [
-        // `key=` (substring match) subsumes the `*key=` variants — `app_key=`,
-        // `apikey=`, `api_key=` all CONTAIN it — so don't re-add those here.
-        "key=",
-        "app_id=",
-        "secret=",
-        "token=",
-        "password=",
-        "pwd=",
-        "auth=",
-        // JSON field shape: `"api_key":"value"` after brace/quote trimming becomes
-        // `api_key":"value`; the `key":` sub-string flags it. `key":` subsumes
-        // `apikey":`, `api_key":`, etc. The whole token is replaced, matching the
-        // same behaviour as the `=` variants above.
-        "key\":",
-        "secret\":",
-        "token\":",
-        "password\":",
-        "auth\":",
-    ]
-    .iter()
-    .any(|marker| lower.contains(marker));
-
-    // Bare IPv4 / host:port — leaks the user's network surroundings. Require an
-    // embedded `.` AND either a trailing `:<digits>` port or an all-numeric dotted
-    // IPv4, so `429:`, `12:34` timestamps, and plain integers stay untouched.
-    let is_host_port = trimmed.contains('.') && {
-        let segs: Vec<&str> = trimmed.split('.').filter(|s| !s.is_empty()).collect();
-        let dotted_ipv4 = segs.len() == 4
-            && segs
-                .iter()
-                .all(|seg| seg.chars().all(|c| c.is_ascii_digit()));
-        let host_with_port = trimmed.rsplit_once(':').is_some_and(|(host, port)| {
-            host.contains('.') && !port.is_empty() && port.chars().all(|c| c.is_ascii_digit())
-        });
-        dotted_ipv4 || host_with_port
-    };
-
-    // Email address: `local@domain.tld` — common in crash logs that include
-    // contact profile data, apply-email generation output, or error context from
-    // profile imports. Require a non-empty local part and a domain bearing a `.`
-    // so bare `@` symbols and TLD-only fragments are left untouched.
-    let is_email = trimmed
-        .split_once('@')
-        .is_some_and(|(local, domain)| !local.is_empty() && domain.contains('.'));
-
-    if is_url {
-        token.replace(trimmed, "<url-redacted>")
-    } else if is_credential {
-        token.replace(trimmed, "<credential-redacted>")
-    } else if is_windows_path || is_unix_path || is_homeish_path {
-        token.replace(trimmed, "<path-redacted>")
-    } else if is_host_port {
-        token.replace(trimmed, "<host-redacted>")
-    } else if is_email {
-        token.replace(trimmed, "<email-redacted>")
-    } else {
-        token.to_string()
-    }
-}
+/// Reason redaction lives in the L0 `observability` module (the L1 board-health
+/// store needs it too, and may not reach up into L2).
+use crate::observability::sanitize_reason;
 
 /// Turn the per-board scrape summaries into a single human-readable reason string
 /// explaining why a run may have come up short — `"<board>: <reason>"` for each
